# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PSA Backend - All Projects",
    "version": "1.0.0",
    "description": "Every endpoint of every project backend in this repository. Operations are tagged and prefixed with the project they belong to."
  },
  "paths": {
...
    "/membership-analytics/api/v1/pulse/constituency/{name}": {
      "get": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Constituency",
        "description": "Booth-level drill-down for one constituency. Heavy on first hit (~tens of\nseconds) \u2014 then materialized into pulse_trend_* tables, instant after.",
        "operationId": "Membership_Analytics__constituency_api_v1_pulse_constituency__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/pulse/constituencies/warm": {
      "post": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Warm Constituencies",
        "description": "Kick off materialization of ALL constituencies in the background. Long-running\n(~minutes per dozen seats); best off-peak. Poll /constituencies/warm/status.",
        "operationId": "Membership_Analytics__warm_constituencies_api_v1_pulse_constituencies_warm_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          },
          {
            "name": "x-refresh-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Refresh-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/membership-analytics/api/v1/pulse/survey": {
      "get": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Survey Intelligence",
        "description": "Survey Intelligence snapshot for an agency (lean \u2014 per-AC vote/margin only).",
        "operationId": "Membership_Analytics__survey_intelligence_api_v1_pulse_survey_get",
        "parameters": [
          {
            "name": "agency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Agency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/membership-analytics/api/v1/pulse/survey/ac": {
      "get": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Survey Ac Report",
        "description": "Full materialized survey blocks for one constituency (AC Report) \u2014 instant snapshot lookup.",
        "operationId": "Membership_Analytics__survey_ac_report_api_v1_pulse_survey_ac_get",
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Name"
            }
          },
          {
            "name": "agency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Agency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/membership-analytics/api/v1/pulse/survey/refresh": {
      "post": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Survey Refresh",
        "operationId": "Membership_Analytics__survey_refresh_api_v1_pulse_survey_refresh_post",
        "parameters": [
          {
            "name": "x-refresh-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Refresh-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/membership-analytics/api/v1/pulse/panel": {
      "get": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Panel",
        "description": "Dynamic panel/volatility for a chosen survey set, e.g. ?sids=19,20,26,28,31.\nPanel = phones present in ALL selected surveys. Computed in-memory (<1s).",
        "operationId": "Membership_Analytics__panel_api_v1_pulse_panel_get",
        "parameters": [
          {
            "name": "sids",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Sids"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/membership-analytics/api/v1/pulse/refresh": {
      "post": {
        "tags": [
          "Membership Analytics - Pulse Trend"
        ],
        "summary": "[Membership Analytics] Refresh",
        "operationId": "Membership_Analytics__refresh_api_v1_pulse_refresh_post",
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          },
          {
            "name": "x-refresh-token",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "title": "X-Refresh-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__app__schemas__nominated_schema__ApiResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Membership_Analytics__HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Membership_Analytics__HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/Membership_Analytics__ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Membership_Analytics__ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "Membership_Analytics__app__schemas__nominated_schema__ApiResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "data": {
            "title": "Data"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "OK"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "ApiResponse"
      },
...
    }
  },
...
}